Add abstract BatchSubmitter
