+++ packages/batch-submitter/src/batch-submitter/index.ts
@@ -1,0 +1,2 @@
+export * from './batch-submitter'
+export * from './tx-batch-submitter'
